+++ src/huntsman/pocs/focuser/birger.py
@@ -1,8 +1,9 @@
-from huntsman.pocs.focuser.serial import HuntsmanSerialFocuser
 from panoptes.pocs.focuser.birger import Focuser as BirgerFocuser
+from huntsman.pocs.focuser import HuntsmanFocuser
 
 
-class Focuser(BirgerFocuser, HuntsmanSerialFocuser):
+class Focuser(BirgerFocuser, HuntsmanFocuser):
+    """ Override class to use methods in HuntsmanFocuser. """
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
